#25 Refactor moving epic actions to actionCreators.js
