+++ src/actionCreators.js
@@ -1,31 +1,40 @@
 import actions from './store/actions';
 
-function updateLoanValue(value) {
-	return {
-		type: actions.UPDATE_LOAN,
-		value,
-	};
-}
+const updateLoanValue = value => ({
+	type: actions.UPDATE_LOAN,
+	value,
+});
 
-function updateMonthsValue(value) {
-	return {
-		type: actions.UPDATE_MONTHS,
-		value,
-	};
-}
+const updateMonthsValue = value => ({
+	type: actions.UPDATE_MONTHS,
+	value,
+});
 
-function updateInsuranceValue(value) {
-	return {
-		type: actions.UPDATE_WITH_INSURANCE,
-		value,
-	};
-}
+const updateInsuranceValue = value => ({
+	type: actions.UPDATE_WITH_INSURANCE,
+	value,
+});
 
-function updateSummaries(payload) {
-	return {
-		type: actions.UPDATE_SUMMARY_FIELDS,
-		payload,
-	};
-}
+const updateSummaries = payload => ({
+	type: actions.UPDATE_SUMMARY_FIELDS,
+	payload,
+});
 
-export { updateLoanValue, updateMonthsValue, updateInsuranceValue, updateSummaries };
+const requestLoanCalc = (amount, time, insurance) => ({
+	type: actions.CALC_LOAN_SEND,
+	payload: { amount, time, insurance },
+});
+
+const requestLoanCalcDebounce = () => ({ type: actions.CALC_LOAN_SEND_DEBOUNCE });
+
+const cancelRequest = () => ({ type: actions.QUERY_CANCELLED, payload: {} });
+
+export {
+	updateLoanValue,
+	updateMonthsValue,
+	updateInsuranceValue,
+	updateSummaries,
+	requestLoanCalc,
+	requestLoanCalcDebounce,
+	cancelRequest,
+};
