+++ src/store/epics/calcLoanDebounce.js
@@ -5,14 +5,9 @@ import { updateSummaries } from '../../actionCreators';
 import { Requests } from '../../utils';
 import actions from '../actions';
 
-const { CALC_LOAN_SEND_DEBOUNCE, QUERY_CANCELLED } = actions;
-
-export const requestLoanCalcDebounce = () => ({ type: CALC_LOAN_SEND_DEBOUNCE });
-export const cancelRequest = () => ({ type: QUERY_CANCELLED, payload: {} });
-
 const calculateLoanDebounceEpic = (action$, state$) =>
 	action$.pipe(
-		ofType(CALC_LOAN_SEND_DEBOUNCE),
+		ofType(actions.CALC_LOAN_SEND_DEBOUNCE),
 		debounceTime(1000),
 		withLatestFrom(state$),
 		switchMap(([, state]) =>
@@ -20,7 +15,7 @@ const calculateLoanDebounceEpic = (action$, state$) =>
 				.getJSON(Requests.adjustLoan(state.borrow.value, state.months.value, state.insurance.withi))
 				.pipe(
 					map(response => updateSummaries(response)),
-					takeUntil(action$.ofType(QUERY_CANCELLED))
+					takeUntil(action$.ofType(actions.QUERY_CANCELLED))
 				)
 		)
 	);
